+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Bumped `crypto-bigint` to 0.5.0-pre.2 with the corresponding changes in the API. (#[13])
+- `Sieve::new()` now does not panic; if the parameters imply an empty output, that's what the iterator generates. (#[14])
 
 
 ### Added
@@ -19,11 +20,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Added a `gcd(Q, n) == 1` check to the Lucas test. (#[11])
+- Added a simple 3 mod 4 check for safe primes, speeding up their generation. (#[14])
+- Fixed multiple corner cases in `Sieve` for small ranges. (#[14])
 
 
 [#11]: https://github.com/nucypher/rust-umbral/pull/11
 [#12]: https://github.com/nucypher/rust-umbral/pull/12
 [#13]: https://github.com/nucypher/rust-umbral/pull/13
+[#14]: https://github.com/nucypher/rust-umbral/pull/14
 
 
 ## [0.1.0] - 2023-01-20