+++ CHANGELOG.md
@@ -4,7 +4,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [0.2.0] - 2023-03-06
 
 ### Changed
 
@@ -37,5 +37,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release.
 
 
-[Unreleased]: https://github.com/entropyxyz/crypto-primes/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/entropyxyz/crypto-primes/compare/v0.2.0...HEAD
 [0.1.0]: https://github.com/nucypher/rust-umbral/releases/tag/v0.1.0
+[0.2.0]: https://github.com/nucypher/rust-umbral/releases/tag/v0.2.0